20240516 qna게시판 프론트 반영, diary게시판 좋아요/댓글삭제권한/게시글삭제 시 댓글삭제

+++ src/main/java/com/example/demo/comment/service/dBoardCommentServiceImpl.java
@@ -3,7 +3,9 @@ package com.example.demo.comment.service;
 import java.util.ArrayList;
 import java.util.List;
 
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Service;
 
 import com.example.demo.comment.dto.CommentDTO;
@@ -16,6 +18,9 @@ public class dBoardCommentServiceImpl implements dBoardCommentService {
 
     @Autowired
     private dBoardCommentRepository repository;
+
+    @Autowired
+    private AuthenticationFacade authenticationFacade;
 
     @Override
     public int register(CommentDTO dto) {
@@ -30,7 +35,7 @@ public class dBoardCommentServiceImpl implements dBoardCommentService {
 
         Diary board = Diary.builder().no(boardNo).build(); // 게시물 조회
 
-        List<dBoardComment> entityList = repository.findByBoard(board);
+        List<dBoardComment> entityList = repository.findByDiary(board);
 
         List<CommentDTO> dtoList = new ArrayList<>();
 
@@ -44,8 +49,18 @@ public class dBoardCommentServiceImpl implements dBoardCommentService {
 
     @Override
     public void remove(int no) {
+        dBoardComment comment = repository.findById(no)
+                .orElseThrow(()-> new EntityNotFoundException("해당 댓글을 찾지 못했습니다. "));
 
-        repository.deleteById(no);
+        Authentication authentication = authenticationFacade.getAuthentication();   // 현재 로그인한 사용자 확인
+
+        if(!comment.getWriter().equals(authentication.getName())){
+            System.out.println("해당 댓글은 삭제 불가능합니다. (접근 불가능)");
+        }else{
+            repository.deleteById(no);
+        }
     }
+
+
 
 }
